requirements.txt
# 기본 데이터 처리
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
Pillow>=8.3.0
opencv-python>=4.5.0

# 딥러닝 프레임워크
torch>=1.9.0
torchvision>=0.10.0
timm>=0.4.0

# 데이터 증강
albumentations>=1.0.0

# 머신러닝
scikit-learn>=1.0.0

# 이미지 처리
imageio>=2.9.0

# 시각화
plotly>=5.0.0

# 유틸리티
tqdm>=4.62.0
pathlib2>=2.3.0

# 실험 추적
wandb>=0.15.0

# Jupyter 노트북
jupyter>=1.0.0
ipykernel>=6.0.0
